refactor: extract lexer creation logic into LexerUtil interface
